refactor: moved shared logic into base class
